pyproject.toml
[project]
name = "ClayCode"
version = "0.1.1"
description = "A python package for building clay models and analysing MD simulation data of clay systems"
authors = [{ name = "Hannah Pollak", email = "h.pollak@sms.ed.ac.uk" }]
license = { file = "LICENSE" }
readme = "README.md"
requires-python = ">=3.9"
dependencies = ['matplotlib>=3.6.2',
    'mdanalysis==2.5',
    'numpy>=1.23.5',
    'pandas~=1.5.2',
    'pyyaml>=6',
    'scipy>=1.9.3',
    'tqdm>=4.43.0',
    'parmed>=4.1',
    'nocasedict',
    'zarr~=2.16',
    'dask~=2023.10',
    'caseless-dictionary>=1',
    'importlib_resources>=6.0',
    'numba>=0.57',
    'pytest',
    "nglview>=3.0",
    ]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: BSD License",
    "Operating System :: OS Independent"
]

[project.scripts]
ClayCode = "ClayCode:__main__.run"

[tool.setuptools.packages.find]
namespaces = true
where=['package', 'tests']

[tool.setuptools]
include-package-data = true

[tool.setuptools.package-data]
"ClayCode.data" = ['*/FF/*.ff/*.itp',
    '*/FF/*.itp', '*/MDP/*.mdp', '*/UCS/*/*.itp', '*/UCS/*/*.gro',
    '*/UCS/*/*.top', '*/CLAYS/*/*/*/*.gro', '*/CLAYS/*/*/*/*.top',
    '*/UCS/*.csv', '*/UCS/*.yaml']
"ClayCode.builder.config" = ['defaults.yaml']
"ClayCodeTests.builder" = ['data/*.yaml', 'data/*.csv']
"ClayCode.siminp.config" = ["defaults.yaml"]
"ClayCode.addmols.config" = ["defaults.yaml", "addtypes.yaml"]

[project.urls]
"Homepage" = "https://github.com/Erastova-group/ClayCode.git"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 79
include = '.pyi?$'
exclude = '''
/(
    .git
    | .hg
    | .mypy_cache
    | .tox
    | .venv
    | _build
    | buck-out
    | build
    | dist
)/
'''
[project.optional-dependencies]
test = [
"pytest>=6.0",
"pytest-xdist>=2.5",
"pytest-cov>=3.0",

]

[tool.pytest.ini_options]
minversion = "6.0"
testpaths = [
"ClayCodeTests",
]
python_files = ["test_*.py", "*_test.py", "test.py"]
